Extract secret syncing into separate file so it can also be used prior to istio install if applicable
